backend/requirements.txt
fastapi
uvicorn[standard]
resend
pytest
httpx
email-validator
pytest-mock
auth0-python
SQLAlchemy
psycopg2-binary
python-jose
python-multipart
langchain
langchain-openai
openai


boto3==1.42.11
botocore==1.42.11
pydantic<2
reportlab
python-docx
